Fix an issue where Firefox’s autofill behavior would not activate properly.
Somehow, perhaps due to interactions with another userscript I am writing which automatically switches from phone number to email, the autofill logic in Firefox would fail to activate even after all of the autocomplete and name attributes were set properly. This change blurs and focuses the input to fix that.